+++ go-app/frontend/wailsjs/go/models.ts
@@ -220,6 +220,7 @@ export namespace main {
 	    }
 	}
 	export class Interop_Profile_Metadata {
+	    Path: string;
 	    UpdatedAt: string;
 	    Warnings: string[];
 	
@@ -229,13 +230,16 @@ export namespace main {
 	
 	    constructor(source: any = {}) {
 	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.Path = source["Path"];
 	        this.UpdatedAt = source["UpdatedAt"];
 	        this.Warnings = source["Warnings"];
 	    }
 	}
 	export class Interop_Profile {
+	    Id: string;
 	    Name: string;
 	    UsbID: string;
+	    AutoSelect?: boolean;
 	    Metadata: Interop_Profile_Metadata;
 	
 	    static createFrom(source: any = {}) {
@@ -244,8 +248,10 @@ export namespace main {
 	
 	    constructor(source: any = {}) {
 	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.Id = source["Id"];
 	        this.Name = source["Name"];
 	        this.UsbID = source["UsbID"];
+	        this.AutoSelect = source["AutoSelect"];
 	        this.Metadata = this.convertValues(source["Metadata"], Interop_Profile_Metadata);
 	    }
 	
@@ -288,6 +294,20 @@ export namespace main {
 	        this.Index = source["Index"];
 	        this.Value = source["Value"];
 	        this.Timestamp = source["Timestamp"];
+	    }
+	}
+	export class Interop_SelectedProfileInfo {
+	    Id: string;
+	    Name: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new Interop_SelectedProfileInfo(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.Id = source["Id"];
+	        this.Name = source["Name"];
 	    }
 	}
 	export class Interop_SharedProfile_Author {